requirements.txt
annotated-types==0.6.0
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.3
cachetools==5.3.2
certifi==2019.3.9
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==42.0.1
dataclasses-json==0.5.8
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2
django-allauth==0.60.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-google-maps==0.13.0
django-leaflet==0.29.0
django-markdown==0.8.4
django-redis==5.4.0
django-rest-framework==0.1.0
django-simple-analytics==1.0.1
django-social-auth==0.7.28
django-user-agents==0.4.0
djangorestframework==3.14.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
Markdown==3.5.1
marshmallow==3.20.2
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pep8==1.7.1
phonepe==1.0.1
phonepe-sdk==1.1.0
Pillow==9.5.0
psycopg==3.1.17
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2018.5
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
urllib3==1.24.2
user-agents==2.2.0
